stabilising autoscaling tests.
