+++ k2s/test/e2e/addons/autoscaling/02-securityenhanced/autoscaling_security_test.go
